Auto-convert TimeTypes to/from arrow time types
These are pretty cheap conversions and makes the Date, Time, Timestamp,
and Duration arrow types must more friendly to work with. I'm not sure
Interval has a real clear mapping, so it's left out for now (it also
mentions that Interval isn't even required for full arrow
compatibility). The last arrow type that doesn't currently map to a
native Julia type is Decimal. That will take a bit more digging because
the spec isn't very detailed on it's exact representation. I plan on
asking the arrow mailing list for more details and we can do that later.