Check if column implements DataAPI.refarray for dict encoding (#12)
* Check if column implements DataAPI.refarray for dict encoding

Fixes #2. The `DataAPI.refarray` and `DataAPI.refpool` are meant to
allow custom array types to signal that they internally store integer
ref indices along with a value pool. This is a great integration with
the arrow format, which has a specific "dict encoded" format that
follows this same format (integer indices w/ value pool). So here we
leverage that by checking if an incoming column implements the interface
and, if so, signal that it should be dict encoded in the arrow format.

* Remove @show and add PooledArrays as test dep

* fix tests

* Fix tests