Ensure elements are converted when indexed from ArrowTypes.ToArrow
Fixes #364. This is an unfortunate bug that was sneaking in under
the guise of "supposedly passing" tests. i.e. we had a test like:

```julia
@test ArrowTypes.ToArrow(Any[1, 3.14]) == [1.0, 3.14]
```
_BUT_, this was misleading, because, in reality, we had:

```julia
x = ArrowTypes.ToArrow(Any[1, 3.14])
@test x === 1
```

i.e. when you indexed an element of `ToArrow`, no actual conversion
was happening. So even though `ToArrow` was doing the work of figuring
out the right concrete promoted type or union type, when you indexed
later, it was just giving you the original element. Which actually
works in a number of cases, for example, if the promoted type was a
`Union` and all elements were one of the Union types, but in this case,
where the promoted case was `Float64` and one element was an integer,
we just got the plain integer out, when in reality, we want that integer
converted to a float.