Pre-allocate buffer
If we let transcode to its own allocation it will allocate a small
vector, start filling it, resize the vector, fill it some more, resize
the vector, etc.

Instead in this commit we pre-allocate a vector of the corect size and
pass it to transcode().

Inspired by https://github.com/apache/arrow-julia/pull/399