Don't treat Vector{UInt8} as Arrow Binary type
Fixes #411. Alternative to #419.

This PR should be compatible with or without the ArrowTypes changes.
I think it's fine to do compat things in Arrow like this as long as
they don't get out of hand and we can eventually remove them as we
bump required ArrowTypes versions and such.

The PR consists of not treating `Vector{UInt8}` as the Arrow Binary
type, which is meant for "binary string"s. Julia has a pretty good
match for that in `Base.CodeUnits`, so instead, we use that to write
Binary and `Vector{UInt8}` is treated as a regular List of Primitive UInt8
type.