Ensure that ArrowTypes.default is defined for Vararg tuples
Fixes #461. This is the other proposal from what I originally suggested.
Though `Tuple{Varag}` is never the concrete type of a value in Julia,
it comes up when dealing with structs with fields that have Vararg
types; not common at all, but it's allowed and happens.

The reflection code here is a little gross, but it seems to be what
Base Julia uses in similar queries to see if a tuple has a vararg
element. I've commented out the Arrow/runtests test for now until
we bump another version and can then uncomment. Unit tests are
added for ArrowTypes in the meantime.