Avoid extending `Type` from Base on Julia 1.12
For some reason, the internal Flatbuf module defines a function called
`Type`. On Julia 1.12, defining a new function by defining methods is
insufficient to distinguish it from defining new constructor methods for
a visible type of the same name. It's now necessary to forward-declare
the function to ensure it takes precedence over the other visible
binding. Without doing so, `Base.Type` is extended in Flatbuf instead.