Fix `OffsetArray` support and clean `CategoricalArray` support
Invalid indices were used with `OffsetArray`s as 1-based indexing was assumed.
Fix this, and always wrap them in a `ToArrow` objet so that they are consistently
turned into 1-based arrays.

This allows dropping special code for `CategoricalArray` in favor of using the
standard `DataAPI.refpool` API combined with the Arrow extension point added to
CategoricalArrays.